health-prediction-service/requirements.txt
fastapi
uvicorn
pydantic-settings
scikit-learn
pandas
numpy
joblib
imbalanced-learn